+++ packages/babel-jest/src/__tests__/index.ts
@@ -5,8 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {Config} from '@jest/types';
 import babelJest from '../index';
+import {makeProjectConfig} from '../../../../TestUtils';
 
 //Mock data for all the tests
 const sourceString = `
@@ -21,15 +21,11 @@ const customMultiply = (obj, mul) => {
 customMultiply({a: 32, dummy: "test"}, 2);
 `;
 
-const mockConfig = {
-  moduleFileExtensions: [],
-};
-
 test(`Returns source string with inline maps when no transformOptions is passed`, () => {
   const result = babelJest.process(
     sourceString,
     'dummy_path.js',
-    (mockConfig as unknown) as Config.ProjectConfig,
+    makeProjectConfig(),
   ) as any;
   expect(typeof result).toBe('object');
   expect(result.code).toBeDefined();