+++ packages/jest-environment/src/index.ts
@@ -8,7 +8,6 @@
 import {Script} from 'vm';
 import {Circus, Config, Global} from '@jest/types';
 import jestMock = require('jest-mock');
-import {ScriptTransformer} from '@jest/transform';
 import {
   JestFakeTimers as LegacyFakeTimers,
   LolexFakeTimers,
@@ -27,6 +26,7 @@ export type EnvironmentContext = Partial<{
 
 // Different Order than https://nodejs.org/api/modules.html#modules_the_module_wrapper , however needs to be in the form [jest-transform]ScriptTransformer accepts
 export type ModuleWrapper = (
+  this: Module['exports'],
   module: Module,
   exports: Module['exports'],
   require: Module['require'],
@@ -43,9 +43,7 @@ export declare class JestEnvironment {
   fakeTimers: LegacyFakeTimers<unknown> | null;
   fakeTimersLolex: LolexFakeTimers | null;
   moduleMocker: jestMock.ModuleMocker | null;
-  runScript(
-    script: Script,
-  ): {[ScriptTransformer.EVAL_RESULT_VARIABLE]: ModuleWrapper} | null;
+  runScript<T = unknown>(script: Script): T | null;
   setup(): Promise<void>;
   teardown(): Promise<void>;
   handleTestEvent?(event: Circus.Event, state: Circus.State): void;